+++ SigProfilerMatrixGenerator/scripts/MutationMatrixGenerator.py
@@ -2203,16 +2203,8 @@ def exome_check(
     initial = True
     udpate_chrom = False
     reference_dir = ref_install.reference_dir()
-    ref_dir = str(reference_dir.path)
 
-    exome_file = (
-        ref_dir
-        + "/references/chromosomes/exome/"
-        + genome
-        + "/"
-        + genome
-        + "_exome.interval_list"
-    )
+    exome_file = str(reference_dir.get_exome_interval_list(genome))
 
     exome_output_path = output_matrix + "vcf_files/" + context + "/"
     if context == "ID":